requirements.txt
streamlit==1.26.0
tensorflow==2.20.0
huggingface_hub==0.18.0
pillow==10.0.0
numpy==1.26.0
